move server_info functionality
